--- biggest CHANGE --- - Authorize everything (sidebar,account,controller,...) - Spilt product AND item - Split user items AND all-items - add change password - add function for return user role_name - perform a lot of bugs --- other facilities as soon :) ---

+++ Controller/do-delete-message.php
@@ -2,11 +2,16 @@
 /* Created By Evan ( Sajad Gholami ) */
 include "__php__.php";
 
+if (!Authorization::check("MessageDelete")){
+    Alert::alerts("متاسفانه شما مجوز حذف پیام های خود را ندارید!");
+    redirect(account("dashboard.php"));
+}
+
 if (isset($_GET['id'])){
     $id = $_GET['id'];
     if (is_numeric($id) && $id > 0){
         $db = new DB();
-        $table = Messages::find("id = {$id} ");
+        $table = Messages::find("id = {$id}");
         if (isset($table[0])){
             Messages::delete( $id );
             Alert::alerts("پیام مورد نظر با موفقیت حذف شد.",null,"success");
